+++ RELEASENOTES.md
@@ -25,10 +25,19 @@ These are changes that have backwards-compatible solutions present and that comp
 
 ---
 
+## Release 0.6.2
+
+**Major Changes**
+
+- Added Navigation component
+
+**Minor Changes**
+
+- Markup updates to Modal, Button, Card, Input, and Page Header
+
 ## Release 0.6.1
 
 - Documentation site updates. *No production code changes in this release*.
-
 
 ## Release 0.6.0